// src/app/service/model/PostDTO.ts
/**
 * SpongeblogSP API
 * Spongeblog blogging API
 *
 * OpenAPI spec version: 1.0.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface PostDTO {
    author?: models.UserInfoDTO;

    category?: models.CategoryDTO;

    /**
     * Post text/html content.
     */
    content?: string;

    /**
     * Creation date of Post in long.
     */
    created?: Date;

    links?: Array<models.Link>;

    /**
     * Modification date of Post in long.
     */
    modified?: Date;

    /**
     * Unique identifier.
     */
    postId?: number;

    /**
     * Display status of the Post.
     */
    status?: PostDTO.StatusEnum;

    /**
     * Tags of Post.
     */
    tags?: Array<models.TagDTO>;

    /**
     * The title of the Blog post.
     */
    title?: string;

}
export namespace PostDTO {
    export enum StatusEnum {
        PUBLIC = <any> 'PUBLIC',
        PRIVATE = <any> 'PRIVATE',
        MAINTENANCE = <any> 'MAINTENANCE',
        DELETED = <any> 'DELETED',
        PAGE_PUBLIC = <any> 'PAGE_PUBLIC',
        PAGE_INACTIVE = <any> 'PAGE_INACTIVE',
    }
}