// src/app/service/model/UserInfoDTO.ts
/**
 * SpongeblogSP API
 * Spongeblog blogging API
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface UserInfoDTO {
    /**
     * Email of User.
     */
    email?: string;

    links?: Array<models.Link>;

    /**
     * Display nickname of the User.
     */
    nickName?: string;

    /**
     * Registration date of User in long.
     */
    registered?: Date;

    /**
     * Unique identifier.
     */
    userId?: number;

}